Errors While Using Prepared Statements
This is the last version of my "sign up" code that uses prepared statements. I got errors in the "invalidUid" function, and in the "createUser" function. I don't know what caused these errors, but I am going to NOT use prepared statements, or as many error handlers, so that I can get some function code started.